Создал функции работы с датой и вынес все функции в отдельный файл
